pizza/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-timeout==5.0.1
boto3==1.38.7
botocore==1.38.7
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
Django==5.2
django-cors-headers==4.7.0
django-redis==5.4.0
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
jmespath==1.0.1
minio==7.2.15
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.22.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==6.1.0
requests==2.32.3
s3transfer==0.12.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
